# api/openapi.yaml
openapi: 3.0.3
info:
  title: Anotify API
  version: "0.1.0"
  description: |
    Anotify 后端 API。静态资源走 /static/*（CDN），动态 API 走 /v1/*（no-store）。
    鉴权两种方式：
      - 会话 Cookie（Passkey 登录后的工作台浏览）
      - Bearer API Key（Agent 上报 / 接收端长连接），Key 仅存哈希，带 scope
servers:
  - url: /v1
...
paths:
...
  /keys:
    get:
      tags: [keys]
      summary: 列出 API Keys（不含明文/哈希）
      responses:
        "200": { description: Key 列表 }
    post:
      tags: [keys]
      summary: 新建 API Key（仅此处返回一次明文）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string }
                scopes:
                  type: array
                  items:
                    {
                      type: string,
                      enum: [notify:send, notify:receive, devices:read],
                    }
      responses:
        "200": { description: 创建的 Key（含明文，仅此一次） }
...
  /devices:
    get:
      tags: [devices]
      summary: 列出接收设备
      responses:
        "200": { description: 设备列表 }
    post:
      tags: [devices]
      summary: 注册/更新推送订阅设备（前端订阅后上报）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceUpsert" }
      responses:
        "200": { description: 已保存 }
  /devices/{id}:
    patch:
      tags: [devices]
      summary: 更新设备（重命名/开关/status 过滤/标签）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                enabled: { type: boolean }
                eventScope: { type: string, enum: [final, all] }
                tags: { type: array, items: { type: string } }
      responses:
        "200": { description: 已更新 }
    delete:
      tags: [devices]
      summary: 解绑设备
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: 已删除 }
...
  /notifications:
    get:
      tags: [notifications]
      summary: 通知历史（分页）
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - {
            name: before,
            in: query,
            schema: { type: string },
            description: 游标(message id),
          }
      responses:
        "200": { description: 历史列表 }
  /notifications/{id}:
    get:
      tags: [notifications]
      summary: 单条通知详情（完整消息 + 真实投递记录）
      description: |
        返回消息的完整字段（broker.Message，camelCase）加 deliveries 数组
        （每台设备/通道的投递状态明细）。仅限消息属主（会话 Cookie）。
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: 消息详情 + deliveries 投递记录 }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: APIKey }
    cookieAuth: { type: apiKey, in: cookie, name: anotify_session }
  schemas:
    DeviceUpsert:
      type: object
      required: [endpoint, keys]
      properties:
        name: { type: string }
        platform: { type: string, enum: [ios, mac, win, android, other] }
        tags: { type: array, items: { type: string } }
        endpoint: { type: string }
        keys:
          type: object
          required: [p256dh, auth]
          properties:
            p256dh: { type: string }
            auth: { type: string }
        userAgent: { type: string }
...
  responses:
    Error:
      description: 错误
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }